Moving searchByManualInput, previousForecastList, and saveForecastList to ViewModel.
