+++ weather-app/src/main/java/rcgonzalezf/org/weather/list/WeatherListViewModel.kt
@@ -1,12 +1,16 @@
 package rcgonzalezf.org.weather.list
 
 import android.app.Application
+import android.content.Context
 import android.location.Address
 import android.location.Geocoder
 import android.util.Log
 import android.widget.Toast
+import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import org.rcgonzalezf.weather.common.ServiceConfig
 import org.rcgonzalezf.weather.common.models.WeatherInfo
 import org.rcgonzalezf.weather.openweather.OpenWeatherApiCallback
@@ -14,6 +18,11 @@ import org.rcgonzalezf.weather.openweather.network.OpenWeatherApiRequestParamete
 import rcgonzalezf.org.weather.R
 import rcgonzalezf.org.weather.common.OnOfflineLoader
 import rcgonzalezf.org.weather.common.ProgressIndicationStateChanger
+import rcgonzalezf.org.weather.utils.WeatherUtils
+import java.io.UnsupportedEncodingException
+import java.net.URLEncoder
+import java.util.concurrent.Executor
+import java.util.concurrent.Executors
 
 class WeatherListViewModel(
         private val openWeatherApiCallback: OpenWeatherApiCallback,
@@ -24,7 +33,11 @@ class WeatherListViewModel(
 
     companion object {
         private val TAG = WeatherListViewModel::class.java.simpleName
+        const val OFFLINE_FILE = "OFFLINE_WEATHER"
+        const val FORECASTS = "FORECASTS"
     }
+
+    private val executor: Executor = Executors.newSingleThreadExecutor()
 
     val cityNameToSearchOnSwipe: MutableLiveData<CharSequence> by lazy {
         MutableLiveData<CharSequence>()
@@ -37,6 +50,19 @@ class WeatherListViewModel(
     val offline: MutableLiveData<Boolean> by lazy {
         MutableLiveData<Boolean>()
     }
+
+    val previousForecastList: List<WeatherInfo>?
+        get() {
+            // TODO Room and LiveData?
+            val sharedPreferences = app.getSharedPreferences(OFFLINE_FILE, 0)
+            val serializedData = sharedPreferences.getString(FORECASTS, null)
+            var storedData: List<WeatherInfo>? = null
+            if (serializedData != null) {
+                storedData = Gson()
+                        .fromJson(serializedData, object : TypeToken<List<WeatherInfo?>?>() {}.type)
+            }
+            return storedData
+        }
 
     fun updateCityNameForSwipeToRefresh(cityName: CharSequence) {
         cityNameToSearchOnSwipe.value = cityName
@@ -55,7 +81,7 @@ class WeatherListViewModel(
                 OpenWeatherApiRequestParameters.OpenWeatherApiRequestBuilder()
                         .withCityName(query)
                         .build(), openWeatherApiCallback)
-        with(app.applicationContext) {
+        with(app) {
             Toast.makeText(this,
                     getString(R.string.searching) + " " + userInput + "...", Toast.LENGTH_SHORT)
                     .show()
@@ -73,5 +99,38 @@ class WeatherListViewModel(
             Log.d(TAG, "error retrieving the cityName with Geocoder")
         }
         return cityName
+    }
+
+    fun saveForecastList(weatherInfoList: List<WeatherInfo>) {
+        executor.execute {
+            // TODO Replace with Room?
+            val prefs = app.getSharedPreferences(OFFLINE_FILE, Context.MODE_PRIVATE)
+            val editor = prefs.edit()
+            editor.putString(FORECASTS, Gson().toJson(weatherInfoList))
+            editor.apply()
+        }
+    }
+
+    @VisibleForTesting
+    fun searchByManualInput(userInput: CharSequence) {
+        // TODO move this to VM, toast with App Context, internet connection as well with App
+        val query: String
+        query = try {
+            URLEncoder.encode(userInput.toString(), "UTF-8")
+        } catch (e: UnsupportedEncodingException) {
+            Log.e(TAG, "Can't encode URL", e)
+            Toast.makeText(app, "${app.getString(R.string.invalid_input)}: $userInput...",
+                    Toast.LENGTH_SHORT).show()
+            return
+        }
+        if (!WeatherUtils.hasInternetConnection(app)) {
+            Toast.makeText(app, app.getString(R.string.no_internet_msg),
+                    Toast.LENGTH_SHORT).show()
+            loadOldData(previousForecastList)
+        } else {
+            // TODO set this in ViewModel
+            offline.value = false
+            searchByQuery(query, userInput)
+        }
     }
 }
